refactor(schemas): split command schemas
